# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Application API for project WineApp",
        "version": "v1"
    },
    "paths": {
...
        "/api/v1/wine": {
            "post": {
                "tags": [
                    "Wine"
                ],
                "summary": "Create UserProject",
                "description": "Create Wine in project with variety",
                "operationId": "createWine",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "project_id",
                                    "wine_variety_id",
                                    "title"
                                ],
                                "properties": {
                                    "project_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "wine_variety_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "title": {
                                        "type": "string",
                                        "example": "Tramín 22"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Response Successfull",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                }
            }
        },
        "/api/v1/wine/{wineId}": {
            "get": {
                "tags": [
                    "Wine"
                ],
                "summary": "Show wine by wineId",
                "description": "Show wine by wineId",
                "operationId": "showWineById",
                "parameters": [
                    {
                        "name": "wineId",
                        "in": "path",
                        "description": "Wine ID",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Response Successfull",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                }
            },
            "put": {
                "tags": [
                    "Wine"
                ],
                "summary": "Update Wine",
                "description": "Update Wine by selected wineId",
                "operationId": "updateWine",
                "parameters": [
                    {
                        "name": "wineId",
                        "in": "path",
                        "description": "Wine ID",
                        "required": true
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [],
                                "properties": {
                                    "wine_variety_id": {
                                        "type": "integer"
                                    },
                                    "title": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Response Successfull",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                }
            }
        },
        "/api/v1/wine/project/{projectId}": {
            "get": {
                "tags": [
                    "Wine"
                ],
                "summary": "Show Wines by Project id",
                "description": "Show Wines by Project id",
                "operationId": "showWineByProjectId",
                "parameters": [
                    {
                        "name": "projectId",
                        "in": "path",
                        "description": "Project ID",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Response Successfull",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                }
            }
        },
...
    }
}